requirements.txt
openai>=1.40.0
pinecone-client>=5.0.0
pydantic>=2.8.0
python-dotenv>=1.0.1
fastapi>=0.112.0
uvicorn>=0.30.0
pypdf>=4.3.1
python-docx>=1.1.2
python-pptx>=0.6.23
pandas>=2.2.2
unstructured>=0.15.0
chardet>=5.2.0
sqlalchemy>=2.0.32
jinja2>=3.1.4
